+++ Tests/SPSC.cpp
@@ -12,7 +12,7 @@ int main()
     SPSCQueue<int> queue(kSize, &alloc);
     size_t sum_expect = 0;
     Thread producer([&]() {
-        for (int i = 0; i < kSize; i++)
+        for (size_t i = 0; i < kSize; i++)
         {
             while (!queue.Push(i));
             sum_expect += i;
